requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
langgraph==0.0.21
langchain==0.0.335
langchain-openai==0.0.2
pydantic==2.4.2
python-dotenv==1.0.0
redis==5.0.1
openai==1.3.5
httpx==0.25.1
sse-starlette==1.6.5
pyngrok==7.0.0
streamlit==1.28.0
pytest==7.4.3
pytest-asyncio==0.21.1
speechrecognition==3.10.0
pydub==0.25.1
nltk==3.8.1
scikit-learn==1.3.2
spacy==3.7.2
pyaudio==0.2.13
python-multipart==0.0.6